+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The local mode now uses the inputs/outputs fields to execute the tasks. The other fields are still kept up-to-date
 - Support for the `transient` flag on tasks outputs (#253)
 
+### Changed
+
+- BREAKING CHANGE: rename `schemas.ComputeTaskOutput` to `schemas.ComputeTaskOutputSpec`
+
 ## [0.32.0](https://github.com/owkin/substra/releases/tag/0.32.0) - 2022-08-09
 
 ### Added